% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.sae}
\alias{print.sae}
\title{Print method for objects of class sae.}
\usage{
\method{print}{sae}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \code{sae}.}

\item{digits}{number of digits to display.}

\item{...}{additional arguments passed to \code{print.default}.}
}
\description{
Print method for objects of class sae.
}
