% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.weights}
\alias{summary.weights}
\title{Summary method for objects of class \code{weights}.}
\usage{
\method{summary}{weights}(object, ...)
}
\arguments{
\item{object}{object of class \code{weights} as returned by function \code{\link{uweights}}.}

\item{...}{not used.}
}
\description{
Summary method for objects of class \code{weights}.
}
\seealso{
\code{\link{uweights}}, \code{\link{plot.weights}}
}
