% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.weights}
\alias{plot.weights}
\title{Plot method for objects of class \code{weights}.}
\usage{
\method{plot}{weights}(
  x,
  log = FALSE,
  breaks = "Scott",
  main = "Distribution of weights",
  xlab = if (log) "log(weight)" else "weight",
  ylab = "frequency",
  col = "cyan",
  ...
)
}
\arguments{
\item{x}{object of class \code{weights} as returned by function \code{\link{uweights}}.}

\item{log}{whether to log-transform the weights.}

\item{breaks}{breaks argument of function \code{\link{hist}}. Default is "Scott".}

\item{main}{main title of plot.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{col}{colour.}

\item{...}{additional arguments passed to \code{\link{hist}}.}
}
\description{
Plot method for objects of class \code{weights}.
}
\seealso{
\code{\link{uweights}}, \code{\link{summary.weights}}
}
