\name{print.sae}
\alias{print.sae}
\title{Print method for objects of class sae.}
\usage{
  \method{print}{sae} (x,
    digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{object of class \code{sae}.}

  \item{digits}{number of digits to display.}

  \item{...}{additional arguments passed to
  \code{print.default}.}
}
\description{
  Print method for objects of class sae.
}

